feat: improve test utilities organization
Create a new `testUtils` folder to enhance the management of test-related
utilities. Move `RecoilObserverValues`, `RecoilObserverOnChange`, and
`renderWithRecoilRoot` into the `testUtils` folder for better organization and
efficiency.
