+++ components/users/user/__test__/user.test.tsx
@@ -1,9 +1,10 @@
-import { getSessionStorage, renderWithRecoilRoot, setSessionStorage } from '@stateLogics/utils';
+import { getSessionStorage, setSessionStorage } from '@stateLogics/utils';
 import { screen, waitFor } from '@testing-library/dom';
 import { mockedUserSession } from '__mock__/next-auth';
 import { Session } from 'next-auth';
 import { SessionProvider } from 'next-auth/react';
 import { User } from '..';
+import { renderWithRecoilRoot } from '@stateLogics/utils/testUtils';
 
 describe('User', () => {
   const renderWithSession = (session: Session | null) => {
