+++ lib/stateLogics/utils/index.tsx
@@ -1,7 +1,4 @@
 import { STORAGE_KEY } from '@constAssertions/storage';
-import { RenderOptions, render } from '@testing-library/react';
-import React, { FC, ReactElement, useEffect } from 'react';
-import { RecoilRoot, RecoilState, atom, useRecoilSnapshot, useRecoilValue } from 'recoil';
 import validator from 'validator';
 
 // Days
@@ -17,12 +14,6 @@ export const queries = (...queries: unknown[]) => queries.filter(Boolean).join('
 export const paths = (...paths: unknown[]) => paths.filter(Boolean).join('') || '';
 
 // test-utils for custom render
-const RecoilRootProvider: FC<{ children: React.ReactNode }> = ({ children }) => {
-  return <RecoilRoot>{children}</RecoilRoot>;
-};
-
-export const renderWithRecoilRoot = (ui: ReactElement, options?: Omit<RenderOptions, 'wrapper'>) =>
-  render(ui, { wrapper: RecoilRootProvider, ...options });
 
 export const fetchWithRetry = async (url: string, options?: {}, retryCount = 3) => {
   const offSession = getSessionStorage(STORAGE_KEY['offSession']);
@@ -93,25 +84,4 @@ export const sanitizeObject = (obj: object) => {
     return [key, typeof value === 'string' ? sanitize(value as string) : value];
   });
   return Object.fromEntries(sanitizedEntries) as typeof obj;
-};
-
-export const RecoilObserverValue = <T,>({ node }: { node: RecoilState<T> | null }) => {
-  const testingOnlyAtom = atom({ key: 'atomForTestingPurposeOnly' });
-  const state = node ? node : testingOnlyAtom;
-  const value = useRecoilSnapshot().getLoadable(state).valueMaybe();
-
-  return <div>{!!value ? 'active' : 'inactive'}</div>;
-};
-
-// recoil test observer: required to observe state change on unit test
-export const RecoilObserverOnChange = <T,>({
-  node,
-  onChange,
-}: {
-  node: RecoilState<T>;
-  onChange: (value: T) => void;
-}) => {
-  const value = useRecoilValue(node);
-  useEffect(() => onChange(value), [onChange, value]);
-  return null;
 };
